Move private functions of OrderController to OrderService
