+++ .idea/misc.xml
@@ -1,10 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="EntryPointsManager">
-    <list size="1">
-      <item index="0" class="java.lang.String" itemvalue="io.micronaut.http.annotation.Post" />
+    <list size="3">
+      <item index="0" class="java.lang.String" itemvalue="io.micronaut.http.annotation.Controller" />
+      <item index="1" class="java.lang.String" itemvalue="io.micronaut.http.annotation.Get" />
+      <item index="2" class="java.lang.String" itemvalue="io.micronaut.http.annotation.Post" />
     </list>
   </component>
   <component name="ExternalStorageConfigurationManager" enabled="true" />
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_17" project-jdk-name="corretto-18" project-jdk-type="JavaSDK" />
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_17" project-jdk-name="zulu-11" project-jdk-type="JavaSDK" />
 </project>
