+++ src/test/kotlin/com/tradingplatform/OrderMatchingTest.kt
@@ -1,8 +1,11 @@
 package com.tradingplatform
 
-import com.tradingplatform.controller.OrderController
 import com.tradingplatform.data.UserRepo
-import com.tradingplatform.model.*
+import com.tradingplatform.model.BuyOrders
+import com.tradingplatform.model.CompletedOrders
+import com.tradingplatform.model.SellOrders
+import com.tradingplatform.model.User
+import com.tradingplatform.services.OrderService
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -24,10 +27,10 @@ class OrderMatchingTest {
         val user1 = User("", "", "", "", "atul_1")
         UserRepo.users[user1.userName] = user1
         user1.wallet.addAmountToFree(100)
-        val objectOfOrderController = OrderController()
+        val orderService = OrderService()
 
         //Act
-        objectOfOrderController.orderHandler(user1.userName, "BUY", 1, 20, "NORMAL")
+        orderService.orderHandler(user1.userName, "BUY", 1, 20, "NORMAL")
 
         //Assert
         Assertions.assertEquals(80, user1.wallet.getFreeAmount())
@@ -45,7 +48,7 @@ class OrderMatchingTest {
         user2.inventory.esopNormal.free = 10
 
         //Act
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user1.userName, "BUY", 5, 20)
         objectOfOrderController.orderHandler(user2.userName, "SELL", 2, 20)
 
@@ -75,7 +78,7 @@ class OrderMatchingTest {
         user2.inventory.esopPerformance.free = 10
 
         //Act
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user2.userName, "SELL", 2, 20, "PERFORMANCE")
         objectOfOrderController.orderHandler(user1.userName, "BUY", 5, 20)
 
@@ -106,7 +109,7 @@ class OrderMatchingTest {
 
 
         //Act
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user1.userName, "SELL", 10, 20)
         objectOfOrderController.orderHandler(user2.userName, "BUY", 5, 20)
 
@@ -135,7 +138,7 @@ class OrderMatchingTest {
         user2.inventory.esopPerformance.free = 10
 
         //Act
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user1.userName, "BUY", 5, 20)
         objectOfOrderController.orderHandler(user2.userName, "SELL", 2, 20, "PERFORMANCE")
 
@@ -164,7 +167,7 @@ class OrderMatchingTest {
 
         //Act
 
-        OrderController().orderHandler(user1.userName, "SELL", 10, 100, "PERFORMANCE")
+        OrderService().orderHandler(user1.userName, "SELL", 10, 100, "PERFORMANCE")
 
         //Assert
         Assertions.assertEquals(40, user1.inventory.esopNormal.free)
@@ -185,7 +188,7 @@ class OrderMatchingTest {
         val user2 = User("", "", "", "", "atul_2")
         UserRepo.users[user2.userName] = user2
         user2.inventory.esopNormal.free = 10
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user2.userName, "SELL", 2, 20)
 
 
@@ -214,7 +217,7 @@ class OrderMatchingTest {
         val user2 = User("", "", "", "", "atul_2")
         UserRepo.users[user2.userName] = user2
         user2.inventory.esopNormal.free = 10
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user1.userName, "BUY", 2, 20)
 
         //Act
@@ -242,7 +245,7 @@ class OrderMatchingTest {
         val user2 = User("", "", "", "", "atul_2")
         UserRepo.users[user2.userName] = user2
         user2.inventory.esopNormal.free = 10
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user1.userName, "BUY", 2, 20)
         objectOfOrderController.orderHandler(user1.userName, "BUY", 2, 30)
 
@@ -271,7 +274,7 @@ class OrderMatchingTest {
         val user2 = User("", "", "", "", "atul_2")
         UserRepo.users[user2.userName] = user2
         user2.inventory.esopNormal.free = 10
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user1.userName, "BUY", 2, 30)
         objectOfOrderController.orderHandler(user1.userName, "BUY", 2, 20)
 
@@ -301,7 +304,7 @@ class OrderMatchingTest {
         val user2 = User("", "", "", "", "atul_2")
         UserRepo.users[user2.userName] = user2
         user2.inventory.esopNormal.free = 10
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user1.userName, "BUY", 2, 20)
         objectOfOrderController.orderHandler(user1.userName, "BUY", 2, 20)
 
@@ -332,7 +335,7 @@ class OrderMatchingTest {
         val user2 = User("", "", "", "", "atul_2")
         UserRepo.users[user2.userName] = user2
         user2.wallet.addAmountToFree(100)
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user1.userName, "SELL", 2, 20)
         objectOfOrderController.orderHandler(user1.userName, "SELL", 2, 30)
 
@@ -363,7 +366,7 @@ class OrderMatchingTest {
         val user2 = User("", "", "", "", "atul_2")
         UserRepo.users[user2.userName] = user2
         user2.wallet.addAmountToFree(100)
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user1.userName, "SELL", 2, 30)
         objectOfOrderController.orderHandler(user1.userName, "SELL", 2, 20)
 
@@ -391,7 +394,7 @@ class OrderMatchingTest {
         val user2 = User("", "", "", "", "atul_2")
         UserRepo.users[user2.userName] = user2
         user2.wallet.addAmountToFree(100)
-        val objectOfOrderController = OrderController()
+        val objectOfOrderController = OrderService()
         objectOfOrderController.orderHandler(user1.userName, "SELL", 2, 20)
         objectOfOrderController.orderHandler(user1.userName, "SELL", 2, 20)
 
@@ -412,4 +415,3 @@ class OrderMatchingTest {
 
 
 }
-
